[iOS] Fix warnings reported with all enabled.

+++ source/ios/AdaptiveCards/AdaptiveCards/AdaptiveCards/ACRCarouselView.mm
@@ -133,7 +133,7 @@
 
 - (void)handleLeftSwipe
 {
-    NSInteger newCarouselPageViewIndex = MIN(self.carouselPageViewIndex+1, self.carouselPageViewList.count-1);
+    NSInteger newCarouselPageViewIndex = MIN(self.carouselPageViewIndex+1, (NSInteger)self.carouselPageViewList.count-1);
     if(newCarouselPageViewIndex == self.carouselPageViewIndex)
     {
         return;
